+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [2023.2.3]
 
 ### Features
+- Import CCv2 `core-customize` as a separate module [#498](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/498)
 - Don't scan CCv2 `js-storefront` and `datahub` sub-folders during project import/refresh [#497](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/497) 
 
 ### `ImpEx` enhancements